feat: unify task forms
