+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 - Docker Compose теперь поднимает MongoDB с логином `admin` и паролем `admin`;
 - Модуль `taskFields.cjs` содержит описание формы задачи для всех клиентов
   `.env.example` обновлён соответствующим URL.
+- Форма создания задачи объединена в компонент `TaskForm`, который используется
+  в веб‑клиенте и мини‑приложении Telegram.
 - В README пояснена переменная `BOT_API_URL` и приведён пример запуска
   локального telegram-bot-api.
 - Скрипт `create_env_from_exports.sh` теперь экранирует значения переменных, что позволяет использовать специальные символы.
