+++ README.md
@@ -25,8 +25,9 @@
   без предупреждений об установке пакета
 - При деплое на Railway Procfile выполняет `scripts/set_bot_commands.sh`,
   синхронизируя единственную команду `/start` в меню бота
-- Форма задачи описана в модуле `bot/shared/taskFields.cjs` и используется
-  как в веб‑приложении, так и в чат‑боте
+ - Форма задачи описана в модуле `bot/shared/taskFields.cjs` и используется
+   как в веб‑приложении, так и в чат‑боте. Модуль содержит полный набор полей
+   для создания и редактирования задач и подходит для мини‑приложения Telegram.
 
 ## Быстрый старт
