+++ ROADMAP.md
@@ -63,3 +63,5 @@
 52. Позволить пользователю редактировать ФИО и телефон в профиле.
 
 53. Вынести поля формы задачи в общий модуль bot/shared/taskFields.cjs.
+54. Объединить формы создания задачи в единый компонент TaskForm и использовать
+    его во всех частях приложения.
